Creating fixture and then broke it
