+++ Tests/BetterBPMCLI.Tests/ModelsFileManagementTests/FixtureCollection.cs
@@ -1,0 +1,5 @@
+﻿namespace BetterBPMCLI.Tests.ModelsFileManagementTests
+{
+    [CollectionDefinition("FixtureCollection")]
+    public class FixtureCollection : ICollectionFixture<FileManagerFixture> { }
+}
